requirements.txt
groq
python-dotenv
pyaudio
pydantic>=1.8.0,<2.0.0

